move map login in useMap hook
